Reorganize the code in node.rs into a couple separate files with different areas of concern to manage growing number of node actor methods
